Closes #288: Props to @lMaxTl.
This is a new commit in the interests in speed and brevity; the original
PR contained numerous extranneous commits. I asked the original
submitter to clean it a little bit, but they never responded, and now
their commit history requires quite a bit of conflict resolution before
it will merge cleanly. I'm taking the initiative to re-do their work in
a single, new commit just to have it done.