Add Awesome Malware list to "Other Lists Online."
This list is an important piece of the full pentest puzzle. While this
list focuses primarily on the pre-exploitation and exploitation phases
of a penetration test, the Awesome Malware list focuses on
post-exploitation tools and resources, which are needed to make any
meaningful use of success with the tools listed on this list.